TGB-5 исправления
Er диаграмма БД
Миграции для бд(init.sql)
Dockerfile изменить на rec.
Docker-com с учетом доп. путей
Разделить Main на файл с Бд
Убрать данные в env
Добавить сохранение изображений
Реализовать логику проверки сообщений раз в неделю
Реализовать отправку сообщений по расписанию
Разложить по папкам
