+++ SQL_Con/ConnectionSQL.py
@@ -1,0 +1,18 @@
+import os
+from dotenv import load_dotenv
+import psycopg2
+
+load_dotenv()
+
+connection = psycopg2.connect(
+        host = os.getenv("DB_HOST"),
+        user = os.getenv("DB_USER"),
+        password = os.getenv("DB_PASSWORD"),
+        database = os.getenv("DB_NAME"),
+        port = 5432
+)
+connection.autocommit = True
+
+with open("./SQL_Con/init.sql", "r") as script:
+    with connection.cursor() as cursor:
+        cursor.execute(script.read())
