+++ docker-compose.yml
@@ -11,6 +11,6 @@ services:
     depends_on:
       - postgres
     build: .
-    env_file: C:\Users\praktikant\Documents\GitHub\tg-bot-listener\environment_variables.env
+    env_file: ./environment_variables.env
     ports:
       - "5000:5000"
