+++ requirements.txt
@@ -1,0 +1,4 @@
+telebot
+dotenv
+psycopg2
+apscheduler
